+++ node-express/apps/app-template/oas.js
@@ -1,14 +1,10 @@
+/*
 const express = require('express')
 const app = express()
 const swaggerUi = require('swagger-ui-express')
-const YAML = require('yamljs')
-const swaggerDocument = YAML.load('./openapi/example.yaml')
-// const swaggerDocument = require('./swagger.json')
-
-const options = { explorer: true }
 
 app.use('/openapi', express.static('openapi'))
-app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument, options))
+app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(require('yamljs').load('./openapi/example.yaml'), { explorer: true }))
 app.get('*', (req, res) => res.send('use /api-docs!'))
 app.listen(8080, () => console.log(`Example app listening at http://localhost:8080`))
 
@@ -28,4 +24,39 @@ app.listen(8080, () => console.log(`Example app listening at http://localhost:80
 //     ]
 //   }
 // }
-// app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(null, options));
+*/
+
+
+const path = require('path');
+const yamljs = require('yamljs');
+const { resolveRefs } = require('json-refs');
+
+/**
+ * Return JSON with resolved references
+ * @param {array | object} root - The structure to find JSON References within (Swagger spec)
+ * @returns {Promise.<JSON>}
+ */
+const multiFileSwagger = (root) => {
+  const options = {
+    filter: ["relative", "remote"],
+    loaderOptions: {
+      processContent: function (res, callback) {
+        callback(null, yamljs.parse(res.text));
+      },
+    },
+  };
+
+  return resolveRefs(root, options).then(
+    function (results) {
+      return results.resolved;
+    },
+    function (err) {
+      console.log(err.stack);
+    }
+  );
+};
+
+
+multiFileSwagger(
+  yamljs.load(path.resolve(__dirname, "./openapi/example.yaml"))
+).then(res => console.log(res))